FormatOps: detection of trailing comma context
